+++ connectToAWSServerViaSSH.sh
@@ -1,0 +1,1 @@
+ssh -i ~/.ssh/linux-demo.pem ec2-user@52.14.154.73