+++ src/main/java/spark/Dao.java
@@ -1,0 +1,9 @@
+package spark;
+
+import java.util.List;
+
+public interface Dao<E> {
+	void insert(E e);
+	
+	List<E> readAll();
+}